requirements.txt
aiogram==3.23.0
aiohttp==3.9.4
python-dotenv==1.0.1
lxml==5.1.0